+++ mycodo/scripts/mycodo_wrapper.c
@@ -122,6 +122,42 @@ int main(int argc, char *argv[]) {
 			strncat(updateScript, "/upgrade_commands.sh upgrade-master", sizeof(updateScript));
 			system(updateScript);
 
+		} else if (strcmp(argv[1], "disable_pigpiod") == 0) {
+
+            char path[255];
+            strncpy(path, argv[0], sizeof(path));
+            dirname(path);
+
+			char updateScript[255];
+			strncpy(updateScript, "/bin/bash ", sizeof(updateScript));
+			strncat(updateScript, path, sizeof(updateScript));
+			strncat(updateScript, "/upgrade_commands.sh disable-pigpiod", sizeof(updateScript));
+			system(updateScript);
+
+		} else if (strcmp(argv[1], "enable_pigpiod_low") == 0) {
+
+            char path[255];
+            strncpy(path, argv[0], sizeof(path));
+            dirname(path);
+
+			char updateScript[255];
+			strncpy(updateScript, "/bin/bash ", sizeof(updateScript));
+			strncat(updateScript, path, sizeof(updateScript));
+			strncat(updateScript, "/upgrade_commands.sh enable-pigpiod-low", sizeof(updateScript));
+			system(updateScript);
+
+		} else if (strcmp(argv[1], "enable_pigpiod_high") == 0) {
+
+            char path[255];
+            strncpy(path, argv[0], sizeof(path));
+            dirname(path);
+
+			char updateScript[255];
+			strncpy(updateScript, "/bin/bash ", sizeof(updateScript));
+			strncat(updateScript, path, sizeof(updateScript));
+			strncat(updateScript, "/upgrade_commands.sh enable-pigpiod-high", sizeof(updateScript));
+			system(updateScript);
+
 		}
 	} else {