+++ mycodo/scripts/upgrade_post.sh
@@ -21,9 +21,9 @@ ${INSTALL_CMD} update-swap-size
 
 ${INSTALL_CMD} setup-virtualenv
 
-${INSTALL_CMD} install-gpiod
+${INSTALL_CMD} install-pigpiod
 
-${INSTALL_CMD} update-gpiod
+${INSTALL_CMD} update-pigpiod
 
 ${INSTALL_CMD} update-wiringpi